[_base.estimator, msm.estimators.msm_estimator]: refactored MSM estimator
